requirements.txt
tensorflow==2.21.0
keras==3.13.2
pandas==3.0.1
numpy==2.4.3
scipy==1.17.1
flask==3.1.3
pillow==12.1.1
requests==2.32.5
scikit-learn
openpyxl==3.1.5
python-dotenv==1.2.2
h5py==3.14.0
absl-py==2.4.0
grpcio==1.78.0
werkzeug==3.1.6
jinja2==3.1.6
click==8.3.1
colorama==0.4.6
rich==14.3.3